# forecasting-api/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Category-Based Forecast API",
    "description": "API for retrieving daily, monthly, and yearly sales forecasts based on product categories with async job processing.",
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "https://api.yourcompany.com/v1",
      "description": "Production Server"
    },
    {
      "url": "https://api.dev.yourcompany.com/v1",
      "description": "Development Server"
    }
  ],
...
  "paths": {
...
    "/forecasts/generate": {
      "post": {
        "summary": "Create Forecast Generation Job",
        "operationId": "createForecastJob",
        "tags": [
          "Jobs"
        ],
        "description": "Creates an async job to generate forecasts for specified parameters when data is not readily available.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForecastJobRequest"
              },
              "examples": {
                "daily_job": {
                  "summary": "Daily forecast generation job",
                  "value": {
                    "category_id": "electronics",
                    "forecast_horizon": "daily",
                    "start_date": "2025-07-01",
                    "end_date": "2025-07-31",
                    "priority": "normal",
                    "webhook_url": "https://yourapp.com/webhooks/forecast-complete"
                  }
                },
                "monthly_job": {
                  "summary": "Monthly forecast generation job",
                  "value": {
                    "category_id": "food_and_beverage",
                    "forecast_horizon": "monthly",
                    "start_date": "2025-01-01",
                    "end_date": "2025-12-01",
                    "priority": "high",
                    "webhook_url": "https://yourapp.com/webhooks/forecast-complete"
                  }
                },
                "yearly_job": {
                  "summary": "Yearly forecast generation job",
                  "value": {
                    "category_id": "all_products",
                    "forecast_horizon": "yearly",
                    "start_date": "2025-01-01",
                    "end_date": "2030-01-01",
                    "priority": "low"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Job created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                },
                "example": {
                  "job_id": "job_67890fghij",
                  "status": "queued",
                  "message": "Forecast generation job created and queued",
                  "estimated_completion_time": "2025-07-06T16:30:00Z",
                  "poll_url": "/jobs/job_67890fghij",
                  "webhook_url": "https://yourapp.com/webhooks/forecast-complete"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid job parameters.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests - Job queue is full.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}": {
      "get": {
        "summary": "Get Job Status",
        "operationId": "getJobStatus",
        "tags": [
          "Jobs"
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "description": "The unique identifier of the job.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "job_12345abcde"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job status retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                },
                "examples": {
                  "job_processing": {
                    "summary": "Job in progress",
                    "value": {
                      "job_id": "job_12345abcde",
                      "status": "processing",
                      "progress": 45,
                      "message": "Generating forecasts for electronics category",
                      "created_at": "2025-07-06T14:30:00Z",
                      "updated_at": "2025-07-06T14:45:00Z",
                      "estimated_completion_time": "2025-07-06T15:00:00Z",
                      "result_url": null,
                      "webhook_url": "https://yourapp.com/webhooks/forecast-complete"
                    }
                  },
                  "job_completed": {
                    "summary": "Job completed",
                    "value": {
                      "job_id": "job_12345abcde",
                      "status": "completed",
                      "progress": 100,
                      "message": "Forecast generation completed successfully",
                      "created_at": "2025-07-06T14:30:00Z",
                      "updated_at": "2025-07-06T14:58:00Z",
                      "estimated_completion_time": "2025-07-06T15:00:00Z",
                      "result_url": "/forecasts?category_id=electronics&forecast_horizon=daily&start_date=2025-07-01&end_date=2025-07-31",
                      "webhook_url": "https://yourapp.com/webhooks/forecast-complete"
                    }
                  },
                  "job_failed": {
                    "summary": "Job failed",
                    "value": {
                      "job_id": "job_12345abcde",
                      "status": "failed",
                      "progress": 0,
                      "message": "Forecast generation failed due to insufficient historical data",
                      "created_at": "2025-07-06T14:30:00Z",
                      "updated_at": "2025-07-06T14:35:00Z",
                      "estimated_completion_time": null,
                      "result_url": null,
                      "webhook_url": "https://yourapp.com/webhooks/forecast-complete",
                      "error_details": {
                        "code": "INSUFFICIENT_DATA",
                        "message": "Not enough historical data for reliable forecasting"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Job not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}/cancel": {
      "post": {
        "summary": "Cancel Job",
        "operationId": "cancelJob",
        "tags": [
          "Jobs"
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "description": "The unique identifier of the job to cancel.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "job_12345abcde"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job cancelled successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Job cannot be cancelled (e.g., already completed).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Job not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/jobs": {
      "get": {
        "summary": "List Jobs",
        "operationId": "listJobs",
        "tags": [
          "Jobs"
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Filter jobs by status.",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "queued",
                "processing",
                "completed",
                "failed",
                "cancelled"
              ]
            }
          },
          {
            "name": "category_id",
            "in": "query",
            "description": "Filter jobs by category.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number for pagination.",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "Number of jobs per page.",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Jobs retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobListResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ForecastJobRequest": {
        "type": "object",
        "description": "Request body for creating a forecast generation job.",
        "properties": {
          "category_id": {
            "type": "string",
            "description": "The category ID for which to generate forecasts.",
            "example": "electronics"
          },
          "forecast_horizon": {
            "type": "string",
            "enum": [
              "daily",
              "monthly",
              "yearly"
            ],
            "description": "The aggregation level of the forecast.",
            "example": "daily"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "description": "The start date of the forecast period.",
            "example": "2025-07-01"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "description": "The end date of the forecast period.",
            "example": "2025-07-31"
          },
          "priority": {
            "type": "string",
            "enum": [
              "low",
              "normal",
              "high"
            ],
            "description": "Priority level for job processing.",
            "default": "normal",
            "example": "normal"
          },
          "webhook_url": {
            "type": "string",
            "format": "url",
            "description": "Optional webhook URL to notify when job completes.",
            "example": "https://yourapp.com/webhooks/forecast-complete"
          }
        },
        "required": [
          "category_id",
          "forecast_horizon",
          "start_date",
          "end_date"
        ]
      },
      "JobResponse": {
        "type": "object",
        "description": "Response when a job is created.",
        "properties": {
          "job_id": {
            "type": "string",
            "description": "Unique identifier for the created job.",
            "example": "job_12345abcde"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "processing",
              "completed",
              "failed",
              "cancelled"
            ],
            "description": "Current status of the job.",
            "example": "queued"
          },
          "message": {
            "type": "string",
            "description": "Human-readable message about the job status.",
            "example": "Forecast generation job created and queued"
          },
          "estimated_completion_time": {
            "type": "string",
            "format": "date-time",
            "description": "Estimated completion time for the job.",
            "example": "2025-07-06T15:00:00Z"
          },
          "poll_url": {
            "type": "string",
            "format": "url",
            "description": "URL to poll for job status updates.",
            "example": "/jobs/job_12345abcde"
          },
          "webhook_url": {
            "type": "string",
            "format": "url",
            "nullable": true,
            "description": "Webhook URL that will be called when job completes.",
            "example": "https://yourapp.com/webhooks/forecast-complete"
          }
        },
        "required": [
          "job_id",
          "status",
          "message",
          "poll_url"
        ]
      },
      "JobStatusResponse": {
        "type": "object",
        "description": "Detailed job status information.",
        "properties": {
          "job_id": {
            "type": "string",
            "description": "Unique identifier for the job.",
            "example": "job_12345abcde"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "processing",
              "completed",
              "failed",
              "cancelled"
            ],
            "description": "Current status of the job.",
            "example": "processing"
          },
          "progress": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100,
            "description": "Job completion progress as a percentage.",
            "example": 45
          },
          "message": {
            "type": "string",
            "description": "Detailed message about the current job status.",
            "example": "Generating forecasts for electronics category"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the job was created.",
            "example": "2025-07-06T14:30:00Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the job status was last updated.",
            "example": "2025-07-06T14:45:00Z"
          },
          "estimated_completion_time": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Estimated completion time for the job.",
            "example": "2025-07-06T15:00:00Z"
          },
          "result_url": {
            "type": "string",
            "format": "url",
            "nullable": true,
            "description": "URL to access the forecast results (available when job is completed).",
            "example": "/forecasts?category_id=electronics&forecast_horizon=daily&start_date=2025-07-01&end_date=2025-07-31"
          },
          "webhook_url": {
            "type": "string",
            "format": "url",
            "nullable": true,
            "description": "Webhook URL that will be called when job completes.",
            "example": "https://yourapp.com/webhooks/forecast-complete"
          },
          "error_details": {
            "type": "object",
            "nullable": true,
            "description": "Error details if the job failed.",
            "properties": {
              "code": {
                "type": "string",
                "description": "Error code.",
                "example": "INSUFFICIENT_DATA"
              },
              "message": {
                "type": "string",
                "description": "Error message.",
                "example": "Not enough historical data for reliable forecasting"
              }
            }
          }
        },
        "required": [
          "job_id",
          "status",
          "progress",
          "message",
          "created_at",
          "updated_at"
        ]
      },
      "JobListResponse": {
        "type": "object",
        "description": "Response containing a list of jobs.",
        "properties": {
          "jobs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JobStatusResponse"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationInfo"
          }
        },
        "required": [
          "jobs",
          "pagination"
        ]
      },
      "PaginationInfo": {
        "type": "object",
        "description": "Pagination details for the forecast response.",
        "properties": {
          "total_items": {
            "type": "integer",
            "description": "Total number of forecast items available for the query.",
            "example": 31
          },
          "total_pages": {
            "type": "integer",
            "description": "Total number of pages for the query.",
            "example": 1
          },
          "current_page": {
            "type": "integer",
            "description": "The current page number.",
            "example": 1
          },
          "page_size": {
            "type": "integer",
            "description": "The number of items per page.",
            "example": 100
          },
          "has_next_page": {
            "type": "boolean",
            "description": "True if there is a next page, false otherwise.",
            "example": false
          },
          "has_prev_page": {
            "type": "boolean",
            "description": "True if there is a previous page, false otherwise.",
            "example": false
          },
          "next_page_link": {
            "type": "string",
            "format": "url",
            "nullable": true,
            "description": "URL to the next page of results, or null if no next page.",
            "example": null
          },
          "prev_page_link": {
            "type": "string",
            "format": "url",
            "nullable": true,
            "description": "URL to the previous page of results, or null if no previous page.",
            "example": null
          }
        },
        "required": [
          "total_items",
          "total_pages",
          "current_page",
          "page_size",
          "has_next_page",
          "has_prev_page"
        ]
      },
...
      "ErrorDetail": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "description": "The field that caused the error (e.g., 'start_date', 'page_size').",
            "example": "start_date"
          },
          "issue": {
            "type": "string",
            "description": "A description of the issue with the field.",
            "example": "Invalid format"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "description": "Standard error response structure.",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "A specific error code for the type of error.",
                "example": "BAD_REQUEST"
              },
              "message": {
                "type": "string",
                "description": "A human-readable message describing the error.",
                "example": "Invalid date format. Expected YYYY-MM-DD."
              },
              "details": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ErrorDetail"
                },
                "description": "Optional array of specific field errors."
              }
            }
          }
        },
        "required": [
          "error"
        ]
      }
    }
  }
}